fix: Update lib.rs

+++ programs/balanced-dollar/src/lib.rs
@@ -46,7 +46,7 @@ pub mod balanced_dollar {
         ctx: Context<'_, '_, '_, 'info, CrossTransfer<'info>>,
         to: String,
         value: u64,
-        icon_bnusd_value: Option<u64>,
+        icon_bnusd_value: Option<u128>,
         data: Option<Vec<u8>>,
     ) -> Result<u128> {
         instructions::cross_transfer(ctx, to, value,icon_bnusd_value, data)
